+++ packages/koot/core/webpack/config/transform-config-last.js
@@ -1,12 +1,61 @@
-// const webpackBundleAnalyzer = require('webpack-bundle-analyzer')
+const fs = require('fs-extra')
+const path = require('path')
+const webpack = require('webpack')
+const {
+    keyConfigBuildDll, filenameDll, filenameDllManifest,
+    keyConfigOutputPathShouldBe
+} = require('../../../defaults/before-build')
 
 /**
  * Webpack 配置处理 - 最终处理
  * @async
- * @param {Object|Array} config 
+ * @param {Object|Array} config webpack 配置对象
+ * @param {Object} kootBuildConfig
  * @return {Object|Array}
  */
-const transform = async (config) => {
+const transform = async (config, kootBuildConfig = {}) => {
+    const {
+        [keyConfigBuildDll]: createDll = false,
+        dist,
+        webpackDll = [],
+    } = kootBuildConfig
+    const {
+        WEBPACK_BUILD_STAGE: STAGE,
+    } = process.env
+
+    // 生成 DLL 包模式: 本次打包仅生成 DLL 包
+    if (createDll) {
+        const defaults = [
+            'react',
+            'react-dom',
+            'redux',
+            'redux-thunk',
+            'react-redux',
+            'react-router',
+            'react-router-redux',
+            'koot',
+        ]
+        const result = Array.isArray(config) ? { ...config[0] } : { ...config }
+        delete result.watch
+        delete result.watchOptions
+        result.entry = {
+            library: (!Array.isArray(webpackDll) || !webpackDll.length) ? defaults : webpackDll
+        }
+        // console.log('result.entry.library', result.entry.library)
+        result.output = {
+            filename: filenameDll,
+            library: "[name]_[hash]",
+            path: STAGE === 'server' ? path.resolve(dist, 'server') : dist
+        }
+        result.plugins.push(
+            new webpack.DllPlugin({
+                // context: path.resolve(__dirname, '../../../../'),
+                path: path.resolve(result.output.path, filenameDllManifest),
+                name: "[name]_[hash]"
+            })
+        )
+        return validate(result, kootBuildConfig)
+    }
 
     // 数组情况，拆分每项分别处理
     if (Array.isArray(config))
@@ -15,6 +64,15 @@ const transform = async (config) => {
     // copy this
     config = Object.assign({}, config)
 
+    return validate(config, kootBuildConfig)
+}
+
+/**
+ * 生效处理
+ * @param {Object} config 
+ * @returns {Object}
+ */
+const validate = (config, kootBuildConfig) => {
     // try to fix a pm2 bug that will currupt [name] value
     if (typeof config.output === 'object') {
         for (let key in config.output) {
@@ -23,14 +81,85 @@ const transform = async (config) => {
         }
     }
 
-    // remove all undefined from plugins
+    validatePlugins(config, kootBuildConfig)
+    validateModuleRules(config, kootBuildConfig)
+
+    // analyze
+    const isAnalyze = (JSON.parse(process.env.WEBPACK_ANALYZE) || config.analyze) ? true : false
+    if (isAnalyze) {
+        config.output.filename = 'entry.[id].[name].js'
+        config.output.chunkFilename = 'chunk.[id].[name].js'
+        // config.plugins.push(
+        //     new (webpackBundleAnalyzer.BundleAnalyzerPlugin)()
+        // )
+    }
+
+    // custom logic use
+    delete config.__ext
+    delete config.spa
+    delete config.analyzer
+    delete config.htmlPath
+    delete config[keyConfigOutputPathShouldBe]
+
+    return config
+}
+
+/**
+ * 生效处理: plugins
+ * @param {Object} config 
+ * @returns {Object}
+ */
+const validatePlugins = (config, kootBuildConfig = {}) => {
+    const {
+        WEBPACK_BUILD_ENV: ENV,
+        WEBPACK_BUILD_STAGE: STAGE,
+    } = process.env
+
+    // 如果没有 plugins 项，创建空 Array
     if (!Array.isArray(config.plugins)) config.plugins = []
+
+    if (ENV === 'dev' && !kootBuildConfig[keyConfigBuildDll]) {
+        // 如果查有 DLL 结果文件，添加 DllReferencePlugin
+        const file = STAGE === 'server'
+            ? path.resolve(kootBuildConfig.dist, 'server', filenameDllManifest)
+            : path.resolve(kootBuildConfig.dist, filenameDllManifest)
+        if (fs.existsSync(file)) {
+            config.plugins.push(
+                new webpack.DllReferencePlugin({
+                    // context: path.resolve(__dirname, '../../../../'),
+                    // scope: match[1],
+                    manifest: file
+                })
+            )
+        }
+        // const dir = path.resolve(kootBuildConfig.dist, dirDll, STAGE)
+        // if (fs.existsSync(dir)) {
+        //     const regex = new RegExp(filenameDllManifest.replace('[name]', '^(.+)') + '$')
+        //     fs.readdirSync(dir)
+        //         .filter(filename => regex.test(filename))
+        //         .forEach(filename => {
+        //             const match = regex.exec(filename)
+        //             config.plugins.push(
+        //                 new webpack.DllReferencePlugin({
+        //                     context: path.resolve(__dirname, '../../../../'),
+        //                     // scope: match[1],
+        //                     manifest: path.resolve(dir, filename)
+        //                 })
+        //             )
+        //         })
+        // }
+    }
+
+    // 清除 plugins 中的空项和非法项
     config.plugins = config.plugins
         .filter(plugin => (
             typeof plugin !== 'undefined' &&
             plugin !== null
         ))
+}
 
+const validateModuleRules = (config) => {
+    // 删除重复 loader
     if (Array.isArray(config.module.rules)) {
         config.module.rules = removeDuplicateObject(config.module.rules)
     }
@@ -51,25 +180,6 @@ const transform = async (config) => {
         })()
         return list.filter(rule => rule != undefined)
     }
-
-    // analyze
-    const isAnalyze = (JSON.parse(process.env.WEBPACK_ANALYZE) || config.analyze) ? true : false
-    if (isAnalyze) {
-        config.output.filename = 'entry.[id].[name].js'
-        config.output.chunkFilename = 'chunk.[id].[name].js'
-        // config.plugins.push(
-        //     new (webpackBundleAnalyzer.BundleAnalyzerPlugin)()
-        // )
-    }
-
-    // custom logic use
-    delete config.__ext
-    delete config.spa
-    delete config.analyzer
-    delete config.htmlPath
-
-    // no ref obj
-    return config
 }
 
 module.exports = transform