+++ packages/koot/defaults/build-config.js
@@ -12,4 +12,5 @@ module.exports = {
     afterBuild: () => { },
     port: undefined,
     defines: {},
+    webpackDll: []
 }