+++ packages/koot/defaults/dev-request-uri.js
@@ -1,0 +1,6 @@
+/**
+ * [开发模式] 特殊的请求 URI
+ */
+module.exports = {
+    dll: '/__koot_dev_dll'
+}