+++ packages/koot/CHANGELOG.md
@@ -1,18 +1,23 @@
 ## 0.7.0
-**2018-11-07**
+**2018-11-??**
 - 核心
   - 配置项
-    - **新** `css` - CSS 打包相关设置，详情请参见文档的 [配置](https://koot.js.org/config) 章节
+    - **新** `css` - CSS 打包相关设置。详情请参见文档的 [配置](https://koot.js.org/config) 章节
+    - **新** `webpack.dll` - 开发模式下供 `webpack.DllPlugin` 使用。webpack 的监控不会处理这些库/library，以期提高开发模式的打包更新速度。详情请参见文档的 [配置](https://koot.js.org/config) 章节
   - 生产模式
     - 使用 `koot-start` 命令时，如果打包过程发生错误，现在会显示更详细的错误记录
   - 开发模式
     - 现在可以同时启动多个 Koot 项目的开发模式了
+    - 使用 `webpack.DllPlugin` 提高打包更新速度
   - 分析模式
     - 输出的文件名结果现在具有可读性
   - 执行打包时会自动清理 `/logs/webpack-config/` 目录下创建于 2 天前的报告文件
   - `/logs/webpack-config/` 目录下新生成的报告文件中，现在会正确的显示正则表达式
-- React 同构
+- React同构 (`ReactApp`)
   - 现在启动服务器时会对设定的端口进行检查。如果端口被占用，会报告相应的错误
+  - 对于传入完整 `store` 的项目，现在每次请求时会尝试使用全新的 `state`
+- 错误修正
+  - 修复生成了错误的多语言跳转 meta 链接地址的问题
 - 添加依赖包
   - `inquirer`
   - `is-port-reachable`
@@ -24,7 +29,12 @@
     - `copy-webpack-plugin` -> _4.6.0_
     - `webpack` -> _4.25.1_
   - patch
+    - `@babel/core` -> _7.1.5_
+    - `@babel/preset-env` -> _7.1.5_
     - `css-loader` -> _1.0.1_
+    - `fs-extra` -> _7.0.1_
+    - `react` -> _16.6.1_
+    - `react-dom` -> _16.6.1_
     - `react-hot-loader` -> _4.3.12_
     - `sp-css-loader` -> _1.5.3_
 - 其他
@@ -51,6 +61,8 @@
       - 项目自创建的 Redux store。提供该对象时，`redux.combineReducers` 配置项会被忽略
   - 开发模式
     - 启用 `babel-loader` 的缓存，加速热更新的响应速度
+  - Webpack
+    - 写入打包配置记录文件时如果发生错误，现在会忽略该错误，并继续流程
 - React
   - 路由 (router) 移除配置外层包裹的空层级
   - 模板 (`ejs`)